Hello updater!
Moved logic out into own file.
Tries to output gg.cmd output directly instead of just waiting for it to end.
